Use another docker image (#1103)
Try to fix our issues with gitlab ci.

# DISCLAIMER
I don't really know if this will work. I pulled the image locally and ran it, `apksigner` is available, `./gradlew assembleFreeRelease` runs without issues, but I don't know if this will behave differently on gitlab. However, this took a long time to run because it is installing an emulator and sdk tools, don't know why.

Co-authored-by: qwerty287 <ndev@web.de>
Reviewed-on: https://codeberg.org/gitnex/GitNex/pulls/1103
Reviewed-by: M M Arif <mmarif@noreply.codeberg.org>
Co-authored-by: qwerty287 <qwerty287@noreply.codeberg.org>
Co-committed-by: qwerty287 <qwerty287@noreply.codeberg.org>