Add fallback for missing version information in database (#1175)
### Describe what your pull request does and which issue you’re targeting
As described in #1174 the app crashes if a bogus version (empty string) is stored in the database. In case of an empty version information in the database, the fallback should be to gracefully treat this as a version 0. This way every check for a "version greater than" fails.

After all the version information will be updated from the server anyways.
<br>
fixes #1174
<br>
- [X] I carefully read the [contribution guidelines](https://codeberg.org/GitNex/GitNex/src/branch/main/CONTRIBUTING.md).
- [X] I'm following the code standards as defined [here](https://codeberg.org/gitnex/GitNex/wiki/Code-Standards).
- [X] By submitting this pull request, I permit GitNex to license my work under the [GNU General Public License v3](https://codeberg.org/GitNex/GitNex/src/branch/main/LICENSE).

Co-authored-by: Fettlaus <mail@fettlaus.de>
Reviewed-on: https://codeberg.org/gitnex/GitNex/pulls/1175
Reviewed-by: qwerty287 <qwerty287@noreply.codeberg.org>
Co-authored-by: Fettlaus <fettlaus@noreply.codeberg.org>
Co-committed-by: Fettlaus <fettlaus@noreply.codeberg.org>