Port app to M3 (Material You) (#1181)
There are not any functinality changes except more of UI changes which comes with M3. There are still few things left but that are not there yet from Google side.

I will be merging this PR as it does not break any function but attaching a few screenshots if you would like to know how it looks like. Feel free to build from `main` to test it out.

Co-authored-by: M M Arif <mmarif@swatian.com>
Reviewed-on: https://codeberg.org/gitnex/GitNex/pulls/1181