[App Only] Notes (#1207)
Closes #656

It is a simple notes feature where you can add your notes related to your repos, issues etc etc.

For me I sometime came across an issue or thing to fix/add but away from PC. I have notes app, but there are so many things in it and sometime it is lost where I save it. So this will be an easy way to remind me of it when needed.

It can be used as a generic Notes app too.

Co-authored-by: M M Arif <mmarif@swatian.com>
Co-authored-by: qwerty287 <ndev@web.de>
Reviewed-on: https://codeberg.org/gitnex/GitNex/pulls/1207
Reviewed-by: qwerty287 <qwerty287@noreply.codeberg.org>